requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.14.2
playsound==1.2.2
PyAudio==0.2.14
pydantic==2.6.4
pydantic_core==2.16.3
pyobjc==10.2
pyobjc-core==10.2
pyobjc-framework-Accessibility==10.2
pyobjc-framework-Accounts==10.2
pyobjc-framework-AddressBook==10.2
pyobjc-framework-AdServices==10.2
pyobjc-framework-AdSupport==10.2
pyobjc-framework-AppleScriptKit==10.2
pyobjc-framework-AppleScriptObjC==10.2
pyobjc-framework-ApplicationServices==10.2
pyobjc-framework-AppTrackingTransparency==10.2
pyobjc-framework-AudioVideoBridging==10.2
pyobjc-framework-AuthenticationServices==10.2
pyobjc-framework-AutomaticAssessmentConfiguration==10.2
pyobjc-framework-Automator==10.2
pyobjc-framework-AVFoundation==10.2
pyobjc-framework-AVKit==10.2
pyobjc-framework-AVRouting==10.2
pyobjc-framework-BackgroundAssets==10.2
pyobjc-framework-BusinessChat==10.2
pyobjc-framework-CalendarStore==10.2
pyobjc-framework-CallKit==10.2
pyobjc-framework-CFNetwork==10.2
pyobjc-framework-ClassKit==10.2
pyobjc-framework-CloudKit==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Collaboration==10.2
pyobjc-framework-ColorSync==10.2
pyobjc-framework-Contacts==10.2
pyobjc-framework-ContactsUI==10.2
pyobjc-framework-CoreAudio==10.2
pyobjc-framework-CoreAudioKit==10.2
pyobjc-framework-CoreBluetooth==10.2
pyobjc-framework-CoreData==10.2
pyobjc-framework-CoreHaptics==10.2
pyobjc-framework-CoreLocation==10.2
pyobjc-framework-CoreMedia==10.2
pyobjc-framework-CoreMediaIO==10.2
pyobjc-framework-CoreMIDI==10.2
pyobjc-framework-CoreML==10.2
pyobjc-framework-CoreMotion==10.2
pyobjc-framework-CoreServices==10.2
pyobjc-framework-CoreSpotlight==10.2
pyobjc-framework-CoreText==10.2
pyobjc-framework-CoreWLAN==10.2
pyobjc-framework-CryptoTokenKit==10.2
pyobjc-framework-DataDetection==10.2
pyobjc-framework-DeviceCheck==10.2
pyobjc-framework-DictionaryServices==10.2
pyobjc-framework-DiscRecording==10.2
pyobjc-framework-DiscRecordingUI==10.2
pyobjc-framework-DiskArbitration==10.2
pyobjc-framework-DVDPlayback==10.2
pyobjc-framework-EventKit==10.2
pyobjc-framework-ExceptionHandling==10.2
pyobjc-framework-ExecutionPolicy==10.2
pyobjc-framework-ExtensionKit==10.2
pyobjc-framework-ExternalAccessory==10.2
pyobjc-framework-FileProvider==10.2
pyobjc-framework-FileProviderUI==10.2
pyobjc-framework-FinderSync==10.2
pyobjc-framework-FSEvents==10.2
pyobjc-framework-GameCenter==10.2
pyobjc-framework-GameController==10.2
pyobjc-framework-GameKit==10.2
pyobjc-framework-GameplayKit==10.2
pyobjc-framework-HealthKit==10.2
pyobjc-framework-ImageCaptureCore==10.2
pyobjc-framework-InputMethodKit==10.2
pyobjc-framework-InstallerPlugins==10.2
pyobjc-framework-InstantMessage==10.2
pyobjc-framework-Intents==10.2
pyobjc-framework-IntentsUI==10.2
pyobjc-framework-IOBluetooth==10.2
pyobjc-framework-IOBluetoothUI==10.2
pyobjc-framework-IOSurface==10.2
pyobjc-framework-iTunesLibrary==10.2
pyobjc-framework-KernelManagement==10.2
pyobjc-framework-LatentSemanticMapping==10.2
pyobjc-framework-LaunchServices==10.2
pyobjc-framework-libdispatch==10.2
pyobjc-framework-libxpc==10.2
pyobjc-framework-LinkPresentation==10.2
pyobjc-framework-LocalAuthentication==10.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.2
pyobjc-framework-MailKit==10.2
pyobjc-framework-MapKit==10.2
pyobjc-framework-MediaAccessibility==10.2
pyobjc-framework-MediaLibrary==10.2
pyobjc-framework-MediaPlayer==10.2
pyobjc-framework-MediaToolbox==10.2
pyobjc-framework-Metal==10.2
pyobjc-framework-MetalFX==10.2
pyobjc-framework-MetalKit==10.2
pyobjc-framework-MetalPerformanceShaders==10.2
pyobjc-framework-MetalPerformanceShadersGraph==10.2
pyobjc-framework-MetricKit==10.2
pyobjc-framework-MLCompute==10.2
pyobjc-framework-ModelIO==10.2
pyobjc-framework-MultipeerConnectivity==10.2
pyobjc-framework-NaturalLanguage==10.2
pyobjc-framework-NetFS==10.2
pyobjc-framework-Network==10.2
pyobjc-framework-NetworkExtension==10.2
pyobjc-framework-NotificationCenter==10.2
pyobjc-framework-OpenDirectory==10.2
pyobjc-framework-OSAKit==10.2
pyobjc-framework-OSLog==10.2
pyobjc-framework-PassKit==10.2
pyobjc-framework-PencilKit==10.2
pyobjc-framework-PHASE==10.2
pyobjc-framework-Photos==10.2
pyobjc-framework-PhotosUI==10.2
pyobjc-framework-PreferencePanes==10.2
pyobjc-framework-PushKit==10.2
pyobjc-framework-Quartz==10.2
pyobjc-framework-QuickLookThumbnailing==10.2
pyobjc-framework-ReplayKit==10.2
pyobjc-framework-SafariServices==10.2
pyobjc-framework-SafetyKit==10.2
pyobjc-framework-SceneKit==10.2
pyobjc-framework-ScreenCaptureKit==10.2
pyobjc-framework-ScreenSaver==10.2
pyobjc-framework-ScreenTime==10.2
pyobjc-framework-ScriptingBridge==10.2
pyobjc-framework-SearchKit==10.2
pyobjc-framework-Security==10.2
pyobjc-framework-SecurityFoundation==10.2
pyobjc-framework-SecurityInterface==10.2
pyobjc-framework-ServiceManagement==10.2
pyobjc-framework-SharedWithYou==10.2
pyobjc-framework-SharedWithYouCore==10.2
pyobjc-framework-ShazamKit==10.2
pyobjc-framework-Social==10.2
pyobjc-framework-SoundAnalysis==10.2
pyobjc-framework-Speech==10.2
pyobjc-framework-SpriteKit==10.2
pyobjc-framework-StoreKit==10.2
pyobjc-framework-SyncServices==10.2
pyobjc-framework-SystemConfiguration==10.2
pyobjc-framework-SystemExtensions==10.2
pyobjc-framework-ThreadNetwork==10.2
pyobjc-framework-UniformTypeIdentifiers==10.2
pyobjc-framework-UserNotifications==10.2
pyobjc-framework-UserNotificationsUI==10.2
pyobjc-framework-VideoSubscriberAccount==10.2
pyobjc-framework-VideoToolbox==10.2
pyobjc-framework-Virtualization==10.2
pyobjc-framework-Vision==10.2
pyobjc-framework-WebKit==10.2
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.2.0
sniffio==1.3.1
SpeechRecognition==3.10.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1